Texture Shader using Goblin XNA
Terrained texture with smooth gradient for different parts of terrain -
grasses, clouds, mountain, sand.
Zoom in and zoom out to find the rendering effect of the textures to
give a realistic effect. - the more your zoom in the more sharper the
image is shown. The further you move away blurrier it gets.